Implement two-step login support.
It's still a bit messy and relies on a bit too much work for the
consumer of the gem, but it seems to work.

I'm not totally convinced that I've done this perfectly - and I'm a
little hesistant about the whole "encrypt the user's password and send
it back to them" thing but.. as long as our secret_key_base is
sufficiently good right?